# docs/swagger.yaml
definitions:
...
  models.BusinessCategory:
    enum:
    - B2B
    - B2C
    - Non_Profit
    - Government
    - Mixed
    type: string
    x-enum-varnames:
    - BusinessCategoryB2B
    - BusinessCategoryB2C
    - BusinessCategoryNonProfit
    - BusinessCategoryGovernment
    - BusinessCategoryMixed
...
  models.BusinessPhase:
    enum:
    - Startup
    - Growth
    - Mature
    - Exit
    type: string
    x-enum-varnames:
    - BusinessPhaseStartup
    - BusinessPhaseGrowth
    - BusinessPhaseMature
    - BusinessPhaseExit
...
  models.BusinessType:
    enum:
    - Consulting
    - Retail
    - Technology
    - Manufacturing
    - Services
    - Other
    type: string
    x-enum-varnames:
    - BusinessTypeConsulting
    - BusinessTypeRetail
    - BusinessTypeTechnology
    - BusinessTypeManufacturing
    - BusinessTypeServices
    - BusinessTypeOther
...
  models.ProjectMemberRole:
    enum:
    - manager
    - contributor
    - reviewer
    type: string
    x-enum-varnames:
    - ProjectMemberRoleManager
    - ProjectMemberRoleContributor
    - ProjectMemberRoleReviewer
...
  models.ProjectStatus:
    enum:
    - planning
    - active
    - on_hold
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - ProjectStatusPlanning
    - ProjectStatusActive
    - ProjectStatusOnHold
    - ProjectStatusCompleted
    - ProjectStatusCancelled
...
  ports.BusinessResponse:
    properties:
      active:
        type: boolean
      address:
        type: string
      business_category:
        $ref: '#/definitions/models.BusinessCategory'
      business_phase:
        $ref: '#/definitions/models.BusinessPhase'
      business_type:
        $ref: '#/definitions/models.BusinessType'
      city:
        type: string
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone_no:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      operator_user:
        $ref: '#/definitions/ports.UserResponse'
      operator_user_id:
        type: integer
      postal_code:
        type: string
      state:
        type: string
      tagline:
        type: string
      updated_at:
        type: string
      value:
        type: number
      website:
        type: string
    type: object
...
  ports.CreateProjectInput:
    properties:
      business_id:
        type: integer
      description:
        type: string
      managed_by_user_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      project_status:
        $ref: '#/definitions/models.ProjectStatus'
      region_ids:
        items:
          type: string
        type: array
      start_date:
        type: string
      target_end_date:
        type: string
    required:
    - managed_by_user_id
    - name
    - project_status
    type: object
...
  ports.ProjectMemberResponse:
    properties:
      joined_at:
        type: string
      project:
        $ref: '#/definitions/ports.ProjectResponse'
      project_id:
        type: integer
      role:
        $ref: '#/definitions/models.ProjectMemberRole'
      user:
        $ref: '#/definitions/ports.UserResponse'
      user_id:
        type: integer
    type: object
...
  ports.ProjectResponse:
    properties:
      actual_end_date:
        type: string
      business:
        $ref: '#/definitions/ports.BusinessResponse'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      manager:
        $ref: '#/definitions/ports.UserResponse'
      members:
        items:
          $ref: '#/definitions/ports.ProjectMemberResponse'
        type: array
      name:
        type: string
      project_status:
        $ref: '#/definitions/models.ProjectStatus'
      regions:
        items:
          $ref: '#/definitions/ports.RegionResponse'
        type: array
      start_date:
        type: string
      target_end_date:
        type: string
      updated_at:
        type: string
    type: object
...
  ports.RegionResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  ports.UpdateProjectInput:
    properties:
      actual_end_date:
        type: string
      business_id:
        type: integer
      description:
        type: string
      managed_by_user_id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      project_status:
        $ref: '#/definitions/models.ProjectStatus'
      region_ids:
        items:
          type: string
        type: array
      start_date:
        type: string
      target_end_date:
        type: string
    type: object
...
  ports.UserResponse:
    properties:
      active:
        type: boolean
      contact_email:
        type: string
      contact_phone_no:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      login_email:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /projects:
    get:
      description: Retrieves a list of all project records.
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            items:
              $ref: '#/definitions/ports.ProjectResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve projects
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project record. Requires authentication.
      parameters:
      - description: Project creation details (Name, ManagedByUserID, ProjectStatus)
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/ports.CreateProjectInput'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/ports.ProjectResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrManagerNotFound
          schema:
            additionalProperties: true
            type: object
        "409":
          description: ErrProjectNameExists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create New Project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Deletes a project record and all related data (members, regions,
        skills). Only the Project Manager can perform this action.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Project deleted successfully (No Content)
        "400":
          description: Invalid project ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the project manager)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Project
      tags:
      - projects
    get:
      description: Retrieves a specific project record by its unique ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project retrieved successfully
          schema:
            $ref: '#/definitions/ports.ProjectResponse'
        "400":
          description: Invalid project ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates an existing project record. Only the Project Manager can
        perform this action.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/ports.UpdateProjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            $ref: '#/definitions/ports.ProjectResponse'
        "400":
          description: Invalid project ID or request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the project manager)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Project Details
      tags:
      - projects
...
swagger: "2.0"